remove force window reload and prevent continuous auth check when authenticated

+++ src/app/components/RouterProtector/RouterGuard.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { useAutoLogin } from "@/app/hooks/useAutoLogin";
+import { useAppStore } from "@/app/store/appStore";
 import {
   authProtectedRoutes,
   preAuthRenderableRoutes,
@@ -22,20 +23,24 @@ export function AuthRouterGuard({
   const router = useRouter();
   const pathname = usePathname();
   const { error, loading, success } = useAutoLogin();
+  const user = useAppStore((state) => state.user);
 
   const routeType = getRouteType(pathname);
 
   useEffect(() => {
-    if (loading) return;
-
+    if (user) return;
     if (success) {
       router.replace(FEEDS_ROUTE);
-    } else if (!success && routeType === "protected") {
+      return;
+    }
+    if (loading) return;
+
+    if (!success && routeType === "protected") {
       console.error(error);
       sessionStorage.clear();
       router.replace(LOGIN_ROUTE);
     }
-  }, [loading, success, routeType, router, error]);
+  }, [loading, success, routeType, router, error, user]);
 
   if (routeType === "protected") {
     if (loading) return <div>{fallbackElement ?? null}</div>;
